Swap container to component/component to container
- CreateEventPage to component as it is only presenting info
- EventForm to container as it is changing state
- create helper for timezones
